requirements.txt
asgiref==3.2.7
Babel==2.8.0
beautifulsoup4==4.9.0
Django==3.0.5
django-bootstrap4==1.1.1
django-phonenumber-field==4.0.0
django-polymorphic==2.1.2
docxtpl==0.9.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
phonenumbers==8.12.1
psycopg2==2.8.5
python-docx==0.8.10
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
